# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-samplestddev",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — the sample standard deviation of a list of numbers, as an `f64`.",
        "description": "This is a pure orchestrator: it owns the *control flow* but delegates every\ncomputation to its dependencies, exactly as specified:\n\n1. `v = (call srvcs-samplevariance {\"values\": values}).result`;\n2. `result = (call srvcs-sqrt {\"value\": v}).result`.\n\nIt performs no validation of its own — invalid input (e.g. a list too short\nfor a sample variance, or a non-numeric element) is rejected by\n`srvcs-samplevariance` and its `422` is forwarded. If a dependency is\nunreachable it reports itself degraded (`503`).",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleStdDevResponse"
                }
              }
            }
          },
          "422": {
            "description": "a dependency rejected the input (forwarded)"
          },
          "500": {
            "description": "a dependency returned a malformed result"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "values"
        ],
        "properties": {
          "values": {
            "type": "object",
            "description": "The list of numbers (the sample) whose standard deviation to compute."
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
      "SampleStdDevResponse": {
        "type": "object",
        "required": [
          "values",
          "result"
        ],
        "properties": {
          "result": {
            "type": "number",
            "format": "double",
            "description": "The sample standard deviation, as an `f64`."
          },
          "values": {
            "type": "object"
          }
        }
      }
    }
  }
}